no error if setting to state container is already in

+++ lib/ruby/containers/container.rb
@@ -187,6 +187,7 @@ end
 
 def destroy_container
   expire_engine_info
+  return true if read_state == 'nocontainer' 
   return  log_error_mesg('Cannot Destroy a container that is not stopped Please stop first', self) if is_active?
   return false unless @container_api.destroy_container(self)  
   @container_id = '-1'
@@ -195,6 +196,7 @@ end
 
 def unpause_container
   expire_engine_info  
+  return true if read_state == 'running' 
   return log_error_mesg('Can\'t  unpause as no paused', self) unless is_paused?
   return false unless @container_api.unpause_container(self)
   expire_engine_info
@@ -241,12 +243,14 @@ end
 
 def start_container
   expire_engine_info
+  return true if read_state == 'running' 
   return log_error_mesg('Can\'t Start Container as ', self) unless read_state == 'stopped'
   return false unless @container_api.start_container(self)
   expire_engine_info   
 end
 def stop_container
-  expire_engine_info
+  expire_engine_info  
+  return true if read_state == 'stopped' 
   return log_error_mesg('Can\'t Stop Container as ', self) unless read_state == 'running'  
   return log_error_mesg('Api failure to stop container' + @container_api.last_error.to_s, self) unless @container_api.stop_container(self)
   expire_engine_info
@@ -254,6 +258,7 @@ end
 
 def pause_container
   expire_engine_info
+  return true if read_state == 'paused' 
   return log_error_mesg('Can\'t Pause Container as not running', self) unless is_running?
   return false unless @container_api.pause_container(self)  
   expire_engine_info
